+++ src/test/java/org/apiphany/json/jackson/JacksonLibraryTest.java
@@ -1,0 +1,23 @@
+package org.apiphany.json.jackson;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
+
+import org.apiphany.utils.Tests;
+import org.junit.jupiter.api.Test;
+import org.morphix.reflection.Constructors;
+
+/**
+ * Test class for {@link JacksonLibrary}.
+ *
+ * @author Radu Sebastian LAZIN
+ */
+class JacksonLibraryTest {
+
+	@Test
+	void shouldThrowExceptionOnCallingConstructor() {
+		UnsupportedOperationException unsupportedOperationException = Tests.verifyDefaultConstructorThrows(JacksonLibrary.class);
+		assertThat(unsupportedOperationException.getMessage(), equalTo(Constructors.MESSAGE_THIS_CLASS_SHOULD_NOT_BE_INSTANTIATED));
+	}
+
+}