+++ src/test/java/org/apiphany/meters/micrometer/MicrometerLibraryTest.java
@@ -1,0 +1,23 @@
+package org.apiphany.meters.micrometer;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
+
+import org.apiphany.utils.Tests;
+import org.junit.jupiter.api.Test;
+import org.morphix.reflection.Constructors;
+
+/**
+ * Test class for {@link MicrometerLibrary}.
+ *
+ * @author Radu Sebastian LAZIN
+ */
+class MicrometerLibraryTest {
+
+	@Test
+	void shouldThrowExceptionOnCallingConstructor() {
+		UnsupportedOperationException unsupportedOperationException = Tests.verifyDefaultConstructorThrows(MicrometerLibrary.class);
+		assertThat(unsupportedOperationException.getMessage(), equalTo(Constructors.MESSAGE_THIS_CLASS_SHOULD_NOT_BE_INSTANTIATED));
+	}
+
+}